feat: Add player API
