+++ cmd/playerpath/internal/handler/api.go
@@ -1,0 +1,38 @@
+package handler
+
+import (
+	"errors"
+	"fmt"
+	"net/http"
+
+	"github.com/go-playground/validator/v10"
+	"github.com/labstack/echo/v4"
+
+	"github.com/cetteup/playerpath/internal/domain/player"
+)
+
+func (h *Handler) HandleGetPlayer(c echo.Context) error {
+	params := struct {
+		PID int `param:"pid" validate:"gt=0"`
+	}{}
+	if err := c.Bind(&params); err != nil {
+		return echo.NewHTTPError(http.StatusBadRequest).SetInternal(fmt.Errorf("failed to bind request parameters: %w", err))
+	}
+
+	if err := validator.New().StructCtx(c.Request().Context(), params); err != nil {
+		return echo.NewHTTPError(http.StatusBadRequest).SetInternal(fmt.Errorf("invalid parameters: %w", err))
+	}
+
+	p, err := h.repository.FindByPID(c.Request().Context(), params.PID)
+	if err != nil {
+		if errors.Is(err, player.ErrPlayerNotFound) {
+			return echo.NewHTTPError(http.StatusNotFound)
+		}
+		if errors.Is(err, player.ErrMultiplePlayersFound) {
+			return echo.NewHTTPError(http.StatusConflict)
+		}
+		return echo.NewHTTPError(http.StatusInternalServerError).SetInternal(fmt.Errorf("failed to find player: %w", err))
+	}
+
+	return c.JSON(http.StatusOK, EncodePlayer(p))
+}
