+++ internal/utils.go
@@ -1,0 +1,5 @@
+package internal
+
+func ToPointer[T any](value T) *T {
+	return &value
+}
